+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- #1257: Prefer 'probably' over 'maybe' when evaluating audio play support.
+
 ## 4.0.3 (2018-10-22)
 
 - New translations: Arabic, Basque, Czech, French, German, Hungarian, Japanese, Norwegian Bokmål, Polish, Romanian, Spanish